requirements.txt
aws-cdk.core
aws-cdk.aws-lambda
aws-cdk.aws-apigatewayv2
aws-cdk.aws-apigatewayv2-integrations